Improve record OPA.
